Update implementation and plan
